login touch on auth
