+++ lib/middleware/auth.js
@@ -1,13 +1,13 @@
 const url = require('url')
 const got = require('got')
 const passport = require('passport')
+const request = require('request')
 const basic = require('express-basic-auth')
 const OAuth2Strategy = require('passport-oauth2').Strategy
-const jwt = require('jsonwebtoken')
-const encryptor = require('simple-encryptor')(process.env.URL_ENCRYPTION_KEY)
 const Users = require('../models/users')
 const Cooperatives = require('../models/cooperatives')
 const assert = require('../assert')
+const nodebb = require('../nodebb')
 
 /**
  * Extend OAuth2Strategy with an implementation for getting Metry user profile
@@ -57,10 +57,20 @@ exports.initialize = function initialize () {
           }
         }
 
+        console.log('So much for this auth')
+
         /**
          * we need to generate the url
          * and then request it
          */
+
+        var authPath = process.env.FORUM_URL + '/auth/brf?brfauth=' + nodebb.authJWT(user)
+        console.log('tryna connect to: ' + authPath)
+        request(authPath, function (error, response, body) {
+          console.log('error:', error) // Print the error if one occurred
+          console.log('statusCode:', response && response.statusCode) // Print the response status code if a response was received
+          console.log('body:', body) // Print the HTML for the Google homepage.
+        })
 
         /**
          * Fetch cooperative meters
@@ -169,6 +179,8 @@ exports.session = function session () {
   return passport.session()
 }
 
+var tryTouch = true
+
 exports.nodebb = function () {
   return function (req, res, next) {
     if (req.isAuthenticated()) {
@@ -177,18 +189,8 @@ exports.nodebb = function () {
         return next()
       }
 
-      var user = req.user
-      var profile = {
-        metryID: user.metryId,
-        name: user.profile.name,
-        email: user.email
-      }
-
-      var encryptedMessage = encryptor.encrypt(profile)
-      var jwtOptions = { expiresIn: '10h' }
-      var webToken = jwt.sign({msg: encryptedMessage}, process.env.BRFENERGI_SESSION_SECRET, jwtOptions)
-      req.user.nodebbAuthJWT = webToken
-      console.log('user: ' + req.user.nodebbAuthJWT)
+      let authJWT = nodebb.authJWT(req.user)
+      req.user.nodebbAuthJWT = authJWT
     }
     next()
   }
