+++ lib/models/users.js
@@ -24,7 +24,7 @@ const UserSchema = new mongoose.Schema({
   },
   hasBoarded: {
     type: Boolean,
-    defaul: false
+    defaul: false // This is a misspelling, but fixing it leads to crashes. Feel free to try.
   },
   isAdmin: Boolean,
   accessToken: String,
