+++ lib/nodebb.js
@@ -1,0 +1,14 @@
+const jwt = require('jsonwebtoken')
+const encryptor = require('simple-encryptor')(process.env.URL_ENCRYPTION_KEY)
+
+exports.authJWT = function (user) {
+  var profile = {
+    metryID: user.metryId,
+    name: user.profile.name,
+    email: user.email
+  }
+
+  var encryptedMessage = encryptor.encrypt(profile)
+  var jwtOptions = { expiresIn: '10h' }
+  return jwt.sign({msg: encryptedMessage}, process.env.BRFENERGI_SESSION_SECRET, jwtOptions)
+}
