+++ package.json
@@ -106,6 +106,7 @@
     "prismic-element": "^0.7.1",
     "prismic-javascript": "^1.5.0-beta.4",
     "prismic-richtext": "^0.2.7",
+    "request": "^2.88.0",
     "simple-encryptor": "^1.4.0",
     "smoothscroll-polyfill": "^0.4.3",
     "y18n": "^3.2.1"
